# api/api.yaml
openapi: 3.0.3
info:
    title: WebDetect Telegram Webhook API
    description: WebDetect Telegram Webhook API
    version: V1.0
    contact:
        name: RFB
servers:
    - url: http://127.0.0.1:6969
tags:
    - name: Update
paths:
    /update:
        post:
            tags:
                - Update
            summary: Update
            description: Webhook Update
            operationId: handleUpdate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/update'
            responses:
                '200':
                    description: OK 
components:
    schemas:
        update:
            type: object
            properties:
                update_id:
                    type: integer
                    format: int64
                message:
                    $ref: '#/components/schemas/message'
        message:
            type: object
            properties:
                update_id:
                    type: integer
                    format: int64
                chat:
                    $ref: '#/components/schemas/chat'
                message_id:
                    type: integer
                    format: int64
                from:
                    $ref: '#/components/schemas/user'
                text:
                    type: string
        chat:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                last_name:
                    type: string
                first_name:
                    type: string
                username:
                    type: string
        user:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                last_name:
                    type: string
                first_name:
                    type: string
                username:
                    type: string

                